requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
build==0.10.0
CacheControl==0.13.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
crashtest==0.4.1
cryptography==41.0.2
distlib==0.3.7
dulwich==0.21.5
filelock==3.12.2
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.7.0
installer==0.7.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==24.2.0
magic-filter==1.0.11
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
packaging==23.1
pexpect==4.8.0
pipenv==2023.8.26
pkginfo==1.9.6
platformdirs==3.10.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pycparser==2.21
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.0
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.0.post1
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.5.24
typing_extensions==4.7.1
urllib3==1.26.16
virtualenv==20.24.3
yarl==1.9.2
zipp==3.15.0
